requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto==2.49.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-cookies==0.0.2
django-countries==7.5.1
django-ninja==0.22.2
idna==3.4
pkg_resources==0.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
